+++ app/javascript/components/game/controls/GameControls.tsx
@@ -74,6 +74,7 @@ export default function GameControls({ game, callback, update }: GameControlsPro
 
   useEffect(() => {
     if (!game.id) { return }
+    console.log(game.closeNeeded)
     displayActions()
   }, [game, game.lastActionIndex, internalUpdate, update])