+++ app/models/utility/server.rb
@@ -2,6 +2,6 @@
 
 module Utility
   class Server
-    VERSION = "0.49"
+    VERSION = "0.50"
   end
 end